[TypeScript SDK[ remove the usage of gateway in e2e test
